Cleaned up on_active lifecycle handling
We now properly remove registrations when on_active returns, for
whatever reason.  We do not do optimistic removal of registrations when
the receivers list is empty as this causes a pretty nasty disconnect
between end user observations being active and internal ones being
inactive / ready to be activated again.  Most importantly, this violates
the contract we provide for on_active such that for any given path the
caller should not expect concurrent calls to on_active.